+++ pittgoogle/types_.py
@@ -3,11 +3,14 @@
 import datetime
 import importlib.resources
 import logging
+import re
 from pathlib import Path
 
 import fastavro
 import yaml
-from attrs import define, field
+from attrs import converters, define, field, evolve
+
+from . import exceptions
 
 LOGGER = logging.getLogger(__name__)
 PACKAGE_DIR = importlib.resources.files(__package__)
@@ -17,32 +20,46 @@ PACKAGE_DIR = importlib.resources.files(__package__)
 class Schema:
     """Class for an individual schema.
 
-    This class is not intended to be used directly. Use :class:`pittgoogle.Schemas` instead.
+    Do not call this class's constructor directly. Instead, load a schema using the registry
+    :class:`pittgoogle.registry.Schemas`.
+
+    ----
     """
 
-    """The name of the schema."""
+    # String _under_ field definition will cause field to appear as a property in rendered docs.
     name: str = field()
-    """The description of the schema."""
+    """Name of the schema."""
+    origin: str = field()
+    """Pointer to the schema's origin. Typically this is a URL to a repo maintained by the survey."""
     description: str = field()
-    """Name of the `Schema` helper method used to load the schema."""
-    _helper: str = field(default="_local_avsc_helper")
-    """Path where the helper can find the schema."""
+    """A description of the schema."""
+    definition: dict | None = field(default=None)
+    """The schema definition used to serialize and deserialize the alert bytes, if one is required."""
+    schemaless_alert_bytes: bool = field(default=False, converter=converters.to_bool)
+    """Whether the alert bytes are schemaless. If True, a valid `definition` is required to
+    serialize or deserialize the alert packet bytes."""
+    _helper: str = field(default="_local_schema_helper")
+    """Name of the helper method that should be used to load the schema definition."""
     path: Path | None = field(default=None)
-    """Mapping of Pitt-Google's generic field names to survey-specific field names."""
+    """Path where the helper can find the schema."""
+    filter_map: dict = field(factory=dict)
+    """Mapping of the filter name as stored in the alert (often an int) to the common name (often a string)."""
+    # The rest don't need string descriptions because they are explicitly defined as properties below.
+    _survey: str | None = field(default=None)
+    # Don't accept _map as an init arg; we'll load it from a yaml file later.
     _map: dict | None = field(default=None, init=False)
-    """The Avro schema loaded by the helper or None if no Avro schema exists."""
-    avsc: dict | None = field(default=None, init=False)
 
     @classmethod
-    def from_yaml(cls, schema_dict: dict) -> "Schema":
+    def _from_yaml(cls, schema_dict: dict, **evolve_schema_dict) -> "Schema":
+        """Create a `Schema` from an entry in the registry's schemas.yml file."""
         # initialize the class, then let the helper finish up
-        schema = cls(**schema_dict)
+        schema = evolve(cls(**schema_dict), **evolve_schema_dict)
         helper = getattr(cls, schema._helper)
         schema = helper(schema)
         return schema
 
     @staticmethod
-    def _local_avsc_helper(schema: "Schema") -> "Schema":
+    def _local_schema_helper(schema: "Schema") -> "Schema":
         """Resolve `schema.path`. If it points to a valid ".avsc" file, load it into `schema.avsc`."""
         # Resolve the path. If it is not None, this helper expects it to be the path to
         # a ".avsc" file relative to the pittgoogle package directory.
@@ -53,32 +70,54 @@ class Schema:
             (schema.path is None) or (schema.path.suffix != ".avsc") or (not schema.path.is_file())
         )
         if invalid_path:
-            schema.avsc = None
+            schema.definition = None
         else:
-            schema.avsc = fastavro.schema.load_schema(schema.path)
+            schema.definition = fastavro.schema.load_schema(schema.path)
 
         return schema
 
     @staticmethod
-    def _lsst_avsc_helper(schema: "Schema") -> "Schema":
-        """Resolve `schema.path`. If it points to a valid ".avsc" file, load the Avro schema."""
+    def _lsst_schema_helper(schema: "Schema") -> "Schema":
+        """Load the Avro schema definition using the ``lsst.alert.packet`` package.
+
+        Raises:
+            SchemaNotFoundError:
+                If an LSST schema called ``schema.name`` cannot be loaded. An error is raised
+                because the LSST alert bytes are schemaless, so ``schema.definition`` will be
+                required in order to deserialize the alert.
+        """
         import lsst.alert.packet.schema
 
-        major, minor = schema.path.split(".")  # [FIXME]
-        schema.path = lsst.alert.packet.schema.get_schema_path(major, minor)
-        schema.avsc = lsst.alert.packet.schema.Schema.from_file(schema.path)
+        version_msg = f"For valid versions, see {schema.origin}."
+
+        # Parse major and minor versions out of schema.name. Expecting syntax "lsst.v<MAJOR>_<MINOR>.alert".
+        try:
+            major, minor = map(int, re.findall(r"\d+", schema.name))
+        except ValueError:
+            msg = (
+                f"Unable to identify major and minor version. Please use the syntax "
+                "'lsst.v<MAJOR>_<MINOR>.alert', replacing '<MAJOR>' and '<MINOR>' with integers. "
+                f"{version_msg}"
+            )
+            raise exceptions.SchemaNotFoundError(msg)
+
+        schema_dir = Path(lsst.alert.packet.schema.get_schema_path(major, minor))
+        schema.path = schema_dir / f"{schema.name}.avsc"
+
+        try:
+            schema.definition = lsst.alert.packet.schema.Schema.from_file(schema.path).definition
+        except fastavro.repository.SchemaRepositoryError:
+            msg = f"Unable to load the schema. {version_msg}"
+            raise exceptions.SchemaNotFoundError(msg)
 
         return schema
 
     @property
     def survey(self) -> str:
-        """Name of the survey. This is the first block (separated by ".") in the schema's name."""
-        return self.name.split(".")[0]
-
-    @property
-    def definition(self) -> str:
-        """Pointer (e.g., URL) to the survey's schema definition."""
-        return self.map.SURVEY_SCHEMA
+        """Name of the survey. This is usually the first part of the schema's name."""
+        if self._survey is None:
+            self._survey = self.name.split(".")[0]
+        return self._survey
 
     @property
     def map(self) -> dict:
@@ -91,37 +130,23 @@ class Schema:
                 raise ValueError(f"no schema map found for schema name '{self.name}'")
         return self._map
 
-    # @property
-    # def avsc(self) -> Optional[dict]:
-    #     """The Avro schema loaded from the file at `self.path`, or None if a valid file cannot be found."""
-    #     # if the schema has already been loaded, return it
-    #     if self._avsc is not None:
-    #         return self._avsc
-
-    #     # if self.path does not point to an existing avro schema file, return None
-    #     if (self.path is None) or (self.path.suffix != ".avsc") or (not self.path.is_file()):
-    #         return None
-
-    #     # load the schema and return it
-    #     self._avsc = fastavro.schema.load_schema(self.path)
-    #     return self._avsc
-
 
 @define(frozen=True)
 class PubsubMessageLike:
     """Container for an incoming alert.
 
-    This class is not intended to be used directly.
-    Use :class:`pittgoogle.Alert` instead.
+    Do not use this class directly. Use :class:`pittgoogle.alert.Alert` instead.
 
     Purpose:
-    It is convenient for the :class:`pittgoogle.Alert` class to work with a message as a
-    `google.cloud.pubsub_v1.types.PubsubMessage`. However, there are many ways to obtain an alert
+    It is convenient for the `Alert` class to work with a message as a
+    `google.cloud.pubsub_v1.types.PubsubMessage`. However, there are many ways to obtain an `Alert`
     that do not result in a `google.cloud.pubsub_v1.types.PubsubMessage` (e.g., an alert packet
     loaded from disk or an incoming message to a Cloud Functions or Cloud Run module). In those
     cases, this class is used to create an object with the same attributes as a
     `google.cloud.pubsub_v1.types.PubsubMessage`. This object is then assigned to the `msg`
     attribute of the `Alert`.
+
+    ----
     """
 
     data: bytes = field()